Move `TOKEN_SECRET` from gateway -> common config
